Split out the XEP-0004 form handler
